refactor(architecture): move shared domain enums from commands to domain modules
- BranchLengthMode to optimize/params.rs
- RerootMode to clock/find_best_root/params.rs
- MethodAncestral to ancestral/params.rs (new file)
- Delete commands/shared/ module (now empty)
- Update all import sites in commands/{ancestral,clock,timetree}/
