+++ packages/treetime/src/ancestral/mod.rs
@@ -5,6 +5,7 @@ pub(crate) mod gtr_inference;
 #[allow(dead_code)]
 pub(crate) mod gtr_inference_dense;
 pub mod marginal;
+pub mod params;
 
 #[cfg(test)]
 mod __tests__;
