+++ packages/treetime/src/clock/find_best_root/params.rs
@@ -2,6 +2,17 @@ use clap::{Args, ValueEnum};
 use serde::{Deserialize, Serialize};
 use smart_default::SmartDefault;
 
+#[derive(Copy, Debug, Clone, PartialEq, Eq, PartialOrd, Ord, SmartDefault, ValueEnum, Serialize, Deserialize)]
+#[value(rename_all = "kebab-case")]
+pub enum RerootMode {
+  #[default]
+  LeastSquares,
+  MinDev,
+  Oldest,
+  ClockFilter,
+  Mrca,
+}
+
 /// Configuration for branch point optimization methods
 #[derive(Debug, Clone, Serialize, Deserialize, SmartDefault)]
 pub enum BranchPointOptimizationParams {
