+++ packages/treetime/src/commands/mod.rs
@@ -4,5 +4,4 @@ pub mod homoplasy;
 pub mod mugration;
 pub mod optimize;
 pub mod prune;
-pub mod shared;
 pub mod timetree;
