--- packages/treetime/src/commands/shared/mod.rs
@@ -1,1 +1,0 @@
-pub mod args;
