+++ packages/treetime/src/commands/timetree/args.rs
@@ -1,5 +1,7 @@
 use crate::alphabet::alphabet::AlphabetName;
-use crate::commands::shared::args::{BranchLengthMode, MethodAncestral, RerootMode};
+use crate::ancestral::params::MethodAncestral;
+use crate::clock::find_best_root::params::RerootMode;
+use crate::optimize::params::BranchLengthMode;
 use crate::gtr::get_gtr::GtrModelName;
 use crate::seq::gap_fill::GapFill;
 use clap::{Parser, ValueEnum, ValueHint};
